Moved markdown plugin code to own file
